refactor(core): delete legacy cartesian code paths
Closes #816.

Removed (pre-1.0 minor bump):

- analyzeAxisVariance() + its @unpunnyfuns/swatchbook-core/variance
  subpath export. Replaced by Project.varianceByPath, the
  load-time-built ReadonlyMap<string, AxisVarianceResult> consumed
  by the smart emitter, MCP get_axis_variance, and the AxisVariance
  doc block. Read project.varianceByPath.get(path) directly.
- buildJointOverrides() shim (deprecated wrapper around
  probeJointOverrides; zero non-test callers).
- Internal emitCss() — the 200-line cartesian-fan-out CSS emitter,
  replaced by emitAxisProjectedCss() in v0.54. Its dataAttr helper
  inlined into css-axis-projected.ts to drop the cross-file import.
- Internal composeProjectCss() from
  @unpunnyfuns/swatchbook-addon (@internal test-only re-export of
  emitAxisProjectedCss).

AxisVarianceResult + VarianceKind relocated from variance.ts into
types.ts since they're load-bearing for Project.varianceByPath and
the wire-format shape; still exported from the barrel.

emitViaTerrazzo() kept — it's the deliberate "drive Terrazzo plugins
from a loaded Project" API surface; no production callers yet but
the export contract is intentional.

Stale comments referencing the removed symbols cleaned up in
variance-analysis.ts, css-axis-projected.ts, blocks/contexts.ts,
addon/channel-types.ts, and the matching test files. Docs site
(apps/docs/docs/reference/core.mdx) updated to document
project.varianceByPath in place of the removed function.
