+++ packages/core/package.json
@@ -31,10 +31,6 @@
       "types": "./dist/index.d.mts",
       "import": "./dist/index.mjs"
     },
-    "./variance": {
-      "types": "./dist/variance.d.mts",
-      "import": "./dist/variance.mjs"
-    },
     "./fuzzy": {
       "types": "./dist/fuzzy.d.mts",
       "import": "./dist/fuzzy.mjs"
@@ -52,7 +48,7 @@
   ],
   "sideEffects": false,
   "scripts": {
-    "build": "tsdown src/index.ts src/variance.ts src/fuzzy.ts src/resolve-at.ts --format esm --dts --clean --sourcemap",
+    "build": "tsdown src/index.ts src/fuzzy.ts src/resolve-at.ts --format esm --dts --clean --sourcemap",
     "typecheck": "tsc --noEmit",
     "test": "vitest run",
     "test:watch": "vitest",
