--- packages/core/src/css.ts
@@ -1,211 +1,0 @@
-import type { TokenNormalized } from '@terrazzo/parser';
-import { generateShorthand, makeCSSVar, transformCSSValue } from '@terrazzo/token-tools/css';
-import { CHROME_ROLES, CHROME_VAR_PREFIX, DEFAULT_CHROME_MAP } from '#/chrome.ts';
-import type { Axis, Permutation, TokenMap } from '#/types.ts';
-
-/** @internal Addon-internal CSS emitter options. Not part of the public API. */
-export interface EmitCssOptions {
-  /** Override the prefix from project config (default: `config.cssVarPrefix ?? ''`). */
-  prefix?: string;
-  /**
-   * Project axes. Multi-axis projects key per-tuple blocks on compound
-   * attribute selectors
-   * (`[data-<prefix>-mode="Dark"][data-<prefix>-brand="Brand A"] { … }`).
-   * Omitted or single-axis projects collapse to the simpler single-attribute
-   * form `[data-<prefix>-theme="…"]`.
-   */
-  axes?: Axis[];
-  /**
-   * Validated chrome-alias entries from `Project.chrome`. Each `source →
-   * target` pair appends `--<prefix>-<source>: var(--<prefix>-<target>);` to
-   * a trailing `:root` block, so blocks that read the fixed chrome paths
-   * resolve through the alias to the consumer's own token values. Entries
-   * must already be validated; unknown source keys or unresolved targets
-   * should be filtered upstream (see `validateChrome`).
-   */
-  chrome?: Record<string, string>;
-}
-
-/**
- * Emit a single concatenated stylesheet.
- *
- * With axes supplied, emits a `:root` block carrying the default-tuple values
- * plus one block per non-default cartesian tuple keyed on a compound attribute
- * selector matching `Project.axes` order. Every var is redeclared per tuple
- * (flat emission) — see `docs/decisions.md` for the rationale.
- *
- * For single-axis projects (resolver with one modifier, or synthetic `theme`),
- * emission keeps the familiar `[data-theme="…"]` shape.
- *
- * @internal Consumers should not depend on this function. The addon calls
- * it to populate the Storybook virtual module; external consumers driving
- * their own build pipeline should use Terrazzo's CLI against the DTCG
- * sources directly.
- */
-export function emitCss(
-  permutations: Permutation[],
-  permutationsResolved: Record<string, TokenMap>,
-  options: EmitCssOptions = {},
-): string {
-  const prefix = options.prefix ?? '';
-  const varOpts = prefix ? { prefix } : {};
-  const transformAlias = (token: TokenNormalized): string =>
-    makeCSSVar(token.id, { ...varOpts, wrapVar: true });
-
-  const axes = options.axes ?? [];
-  const multiAxis = axes.length > 1;
-  const defaultTuple = buildDefaultTuple(axes);
-  const defaultPermutation =
-    axes.length > 0 ? findPermutationByTuple(permutations, defaultTuple) : undefined;
-
-  const blocks: string[] = [];
-
-  if (multiAxis && defaultPermutation) {
-    const decls = declarationsFor(
-      defaultPermutation,
-      permutationsResolved,
-      varOpts,
-      transformAlias,
-    );
-    if (decls.length > 0) blocks.push(`:root {\n${decls.join('\n')}\n}`);
-  }
-
-  for (const perm of permutations) {
-    if (multiAxis && perm === defaultPermutation) continue;
-    const decls = declarationsFor(perm, permutationsResolved, varOpts, transformAlias);
-    if (decls.length === 0) continue;
-    blocks.push(`${selectorFor(perm, axes, prefix)} {\n${decls.join('\n')}\n}`);
-  }
-
-  const chrome = options.chrome ?? {};
-  const chromeLines: string[] = ['  color-scheme: light dark;'];
-  for (const role of CHROME_ROLES) {
-    const sourceVar = makeCSSVar(role, { prefix: CHROME_VAR_PREFIX });
-    const target = chrome[role];
-    if (target !== undefined) {
-      const targetVar = makeCSSVar(target, { ...varOpts, wrapVar: true });
-      chromeLines.push(`  ${sourceVar}: ${targetVar};`);
-    } else {
-      chromeLines.push(`  ${sourceVar}: ${DEFAULT_CHROME_MAP[role]};`);
-    }
-  }
-  blocks.push(`:root {\n${chromeLines.join('\n')}\n}`);
-
-  return `${blocks.join('\n\n')}\n`;
-}
-
-type VarOpts = Record<string, never> | { prefix: string };
-
-function declarationsFor(
-  perm: Permutation,
-  permutationsResolved: Record<string, TokenMap>,
-  varOpts: VarOpts,
-  transformAlias: (token: TokenNormalized) => string,
-): string[] {
-  const tokens = permutationsResolved[perm.name];
-  if (!tokens) return [];
-  const declarations: string[] = [];
-  for (const [localID, token] of Object.entries(tokens)) {
-    for (const decl of emitTokenDeclarations(
-      localID,
-      token,
-      tokens,
-      perm.input,
-      varOpts,
-      transformAlias,
-    )) {
-      declarations.push(`  ${decl}`);
-    }
-  }
-  return declarations;
-}
-
-/**
- * Compose the selector for a tuple. Multi-axis projects get a compound
- * attribute selector in axis order; single-axis projects keep the familiar
- * `[data-theme="…"]` shape so the single-attribute selector stays recognizable.
- *
- * When `prefix` is set, data-attr names get the same prefix the CSS vars use —
- * e.g. `cssVarPrefix: 'sb'` yields `[data-sb-mode="Dark"]` instead of
- * `[data-mode="Dark"]`. Avoids collisions with third-party libs that claim
- * bare `data-mode` / `data-theme`. Empty prefix keeps the bare form as
- * opt-out.
- */
-function selectorFor(perm: Permutation, axes: Axis[], prefix: string): string {
-  const attr = attrName(prefix);
-  if (axes.length <= 1) {
-    return `[${attr('theme')}="${cssEscape(perm.name)}"]`;
-  }
-  const parts: string[] = [];
-  for (const axis of axes) {
-    const value = perm.input[axis.name];
-    if (value === undefined) continue;
-    parts.push(`[${attr(axis.name)}="${cssEscape(value)}"]`);
-  }
-  return parts.length > 0 ? parts.join('') : `[${attr('theme')}="${cssEscape(perm.name)}"]`;
-}
-
-/**
- * Helper producing a prefixed `data-*` attribute name when `prefix` is set,
- * `data-*` otherwise. Exported so the addon + blocks can stay in lockstep
- * without re-implementing the naming rule.
- */
-export function dataAttr(prefix: string, key: string): string {
-  return prefix ? `data-${prefix}-${key}` : `data-${key}`;
-}
-
-function attrName(prefix: string): (key: string) => string {
-  return (key) => dataAttr(prefix, key);
-}
-
-function buildDefaultTuple(axes: Axis[]): Record<string, string> {
-  const out: Record<string, string> = {};
-  for (const axis of axes) out[axis.name] = axis.default;
-  return out;
-}
-
-function findPermutationByTuple(
-  permutations: Permutation[],
-  tuple: Readonly<Record<string, string>>,
-): Permutation | undefined {
-  const keys = Object.keys(tuple);
-  return permutations.find((perm) => {
-    for (const key of keys) {
-      if (perm.input[key] !== tuple[key]) return false;
-    }
-    return Object.keys(perm.input).length === keys.length;
-  });
-}
-
-function emitTokenDeclarations(
-  localID: string,
-  token: TokenNormalized,
-  tokensSet: TokenMap,
-  permutation: Record<string, string>,
-  varOpts: VarOpts,
-  transformAlias: (token: TokenNormalized) => string,
-): string[] {
-  const varName = makeCSSVar(localID, varOpts);
-
-  const value = transformCSSValue(token, { tokensSet, permutation, transformAlias });
-
-  if (typeof value === 'string') {
-    return [`${varName}: ${value};`];
-  }
-
-  const lines: string[] = [];
-  for (const [subKey, subVal] of Object.entries(value)) {
-    const subName = makeCSSVar(`${localID}.${subKey}`, varOpts);
-    lines.push(`${subName}: ${subVal};`);
-  }
-  const shorthand = generateShorthand({ token, localID });
-  if (shorthand) {
-    lines.push(`${varName}: ${shorthand};`);
-  }
-  return lines;
-}
-
-/** Escape `"` and `\` so attribute values sit safely inside `[attr="…"]`. */
-function cssEscape(name: string): string {
-  return name.replaceAll('\\', '\\\\').replaceAll('"', '\\"');
-}
